# api-docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.SummarySettingsRequest:
    properties:
      default_model:
        minLength: 1
        type: string
    required:
    - default_model
    type: object
  api.SummarySettingsResponse:
    properties:
      default_model:
        type: string
    type: object
  api.SummaryTemplateRequest:
    properties:
      description:
        type: string
      include_speaker_info:
        type: boolean
      model:
        minLength: 1
        type: string
      name:
        minLength: 1
        type: string
      prompt:
        minLength: 1
        type: string
    required:
    - model
    - name
    - prompt
    type: object
...
  models.SummaryTemplate:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      include_speaker_info:
        type: boolean
      model:
        type: string
      name:
        type: string
      prompt:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Audio transcription service using WhisperX
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: aScribe API
  version: "1.0"
paths:
...
  /api/v1/summaries:
    get:
      description: Get all summarization templates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SummaryTemplate'
            type: array
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      - BearerAuth: []
      summary: List summarization templates
      tags:
      - summaries
    post:
      consumes:
      - application/json
      description: Create a new summarization template
      parameters:
      - description: Template payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SummaryTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SummaryTemplate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      - BearerAuth: []
      summary: Create summarization template
      tags:
      - summaries
  /api/v1/summaries/{id}:
    delete:
      description: Delete a summarization template by ID
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete summarization template
      tags:
      - summaries
    get:
      description: Get a summarization template by ID
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SummaryTemplate'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      - BearerAuth: []
      summary: Get summarization template
      tags:
      - summaries
    put:
      consumes:
      - application/json
      description: Update a summarization template by ID
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Template payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SummaryTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SummaryTemplate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      - BearerAuth: []
      summary: Update summarization template
      tags:
      - summaries
  /api/v1/summaries/settings:
    get:
      description: Get global summarization settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SummarySettingsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get summary settings
      tags:
      - summaries
    post:
      consumes:
      - application/json
      description: Create or update global summarization settings
      parameters:
      - description: Settings payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SummarySettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SummarySettingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Save summary settings
      tags:
      - summaries
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"